+++ src/fable-library-ts/Uri.ts
@@ -177,7 +177,13 @@ export class Uri {
   }
 
   get host() {
-    return this.asUrl().host;
+    const host = this.asUrl().host;
+
+    if (host.includes(":")) {
+      return host.split(":")[0];
+    } else {
+      return host;
+    }
   }
 
   get absolutePath() {
@@ -186,6 +192,20 @@ export class Uri {
 
   get query() {
     return this.asUrl().search;
+  }
+
+  get isDefaultPort() {
+    return this.port === 80;
+  }
+
+  get port() {
+    const port = this.asUrl().port;
+
+    if (port === "") {
+      return 80;
+    } else {
+      return parseInt(port);
+    }
   }
 
   get pathAndQuery() {